added analog temp sensor
this change adds an analog temp sensor to tkgpio.py which mimics the same way that the potentiometer was implemented (using mcp adc). Below is a demonstration of how temp sensor could be added to circuit.

"adc": {
       "mcp_chip": 3002,
       "potenciometers": [
       {"x": 175, "y": 375, "name": "Potentiometer on ADC0", "channel": 0}
       ],
       "temp_sensors": [
        {"x": 350, "y": 375, "name": "Temp Sensor on ADC1", "channel": 1}
        ]
    }

This requires that both "potenciometers" and "temp_sensors" be always present, but that if either is not in use then none should be listed within the square brackets. For example, below would implement 0 pots and 1 temp sensor...

"adc": {
       "mcp_chip": 3002,
       "potenciometers": [
       
       ],
       "temp_sensors": [
        {"x": 350, "y": 375, "name": "Temp Sensor on ADC1", "channel": 1}
        ]
    }