improved rgb-led
Tried original wallysami rgb-led version and found the use of alpha to caused a lack of brightness in LED rendering and also resulted in less realistic blending of red-green-blue values. 

This version addresses both of those issues. It uses realistic brightness and realistic colors.

This commit includes slight correction to previous commit.